feat: add product-details basic layout
